Improve api handler

+++ pages/api/post/[postId].ts
@@ -7,35 +7,25 @@ import apiHandler, { MethodHandler } from "../../../util/apiHandler";
 import User, { UserModel } from "../../../models/user";
 
 const methodHandler: MethodHandler = {
-  DELETE(request: NextApiRequest, response: NextApiResponse) {
-    verifyToken(request, response, async (request, response) => {
-      try {
-        const { postId } = request.query;
-        await connectToDatabase();
-        await Post.findOneAndRemove({
-          _id: new Types.ObjectId(postId.toString()),
-        });
-        response.json({ status: "ok", message: "deleted!" });
-      } catch (error) {
-        console.trace(error);
-        response.status(500).send("Internal Exception");
-      }
+  async DELETE(request, response) {
+    const user = verifyToken(request, response);
+    if (!user) return;
+    const { postId } = request.query;
+    await connectToDatabase();
+    await Post.findOneAndRemove({
+      _id: new Types.ObjectId(postId.toString()),
     });
+    response.json({ status: "ok", message: "deleted!" });
   },
-  async GET(request: NextApiRequest, response: NextApiResponse) {
-    try {
-      const { postId } = request.query;
-      await connectToDatabase();
-      const post = (await Post.findById(
-        new Types.ObjectId(postId.toString())
-      )) as PostModel;
-      const user = (await User.findById(post.author_id)) as UserModel;
-      delete user.password;
-      response.json({ post: post, author: user });
-    } catch (error) {
-      console.trace(error);
-      response.status(500).send("Internal Exception");
-    }
+  async GET(request, response) {
+    const { postId } = request.query;
+    await connectToDatabase();
+    const post = (await Post.findById(
+      new Types.ObjectId(postId.toString())
+    )) as PostModel;
+    const user = (await User.findById(post.author_id)) as UserModel;
+    delete user.password;
+    response.json({ post: post, author: user });
   },
 };
 
@@ -43,5 +33,5 @@ export default function handler(
   request: NextApiRequest,
   response: NextApiResponse
 ) {
-  apiHandler(request, response, methodHandler);
+  return apiHandler(request, response, methodHandler);
 }
